requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.0.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
psycopg2==2.8.6
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.13
